ref: init `UserCartModal` component
